requirements.txt
requests==2.32.3
python-telegram-bot==13.7
environs==9.3.3